Support show statement again.
The 'show' statement allows the internal graph structure to be displayed
graphically, without attempting to reduce the graph to a value. This is
for debugging/educational purposes only, and does not work when fundy
has been translated, as the viewing code is not RPython.

TODO: add a nice way of making the same code work on top of CPython and
after translation (by removing the show statement when translating).
Currently you have to edit the grammar to remove the show statement, or
edit asteval.py to make the show statement do something else, in order
to be able to translate.