Automate the translation process with a Makefile.
Includes now having a target spec file for Fundy to pass to
PYPY_DIR/pypy/translator/goal/translate.py, and using the same
source for translation and for running on CPython.

This last point has meant implementing a conditional inclusion
mechanism for the grammar file (#!if ... #!endif), as the show
statement can only be supported while running untranslated at
the moment.