Reorganise the show code a bit. Now shows labels.
Previously, labels for the graphs would only be shown when using an
empty show statement, which displays all things bound to names in the
current context. Now using show on one or more expressions will label
the root of the graph for each expression a textual representation of
the expression.

Consequently, the "show entire context" functionality is now implemented
by asking the context for its entire contents and treating this output
very similarly to input expressions, so Contexts no longer know anything
at all about the dot view code.