Add runtime type information.
No ability to do user-level type checking yet. Type checking of
arguments to builtin functions is now in terms of the Fundy-level type
information now, rather than in terms of the class heirarchy used to
implement Fundy graphs.

There is a fundamental problem, however. If a graph of a function
application is reduced to a type, the node pointer of the application
cell will be reused (necessary for laziness). This means there is now a
new node pointer referring to the type. Consequently, tests of whether
node pointers are contained in a value's type set cannot always give
correct results.

The appropriate fix is to rewrite the whole graph module and get rid of
the pointer/node distinction. Probably geting rid of the class heirarchy
of nodes, making them genuinely overwriteable again. Must think on the
least horrible way to do this.