Type sets see NodePtrs with same node as the same.
See comments in graph.py about the eq and hash methods of NodePtrs.

This should prevent the potential bug discussed in the last commit
message (multiple NodePtrs referring to the same type node not being
considered as matching each other when type checking). Need to construct
a test case for this issue to be sure.