Support layout rules similar to Haskell.
No longer dependent on the return statement for recognising the end of
blocks. Untill now the only form of block has been the def statement,
but that is about to change with typeswitching. Now the parser has a
general idea of "paragraphs" of code that are indented more than the
line they start on. A new block (that may have subblocks) is signalled
by the BEGIN token (currently a singal colon). A block is significant
because its sub-parts are all separated by the TERM token, wheras other
indented paragraphs are treated as if there was only whitespace
separating them.