Switch between translated and untranslated modes.
It is now possible (not by directly running interactive.py, but by using
an interactive Python session) to run Fundy on top of CPython with all
the definitions as if it were being translated. It is even possible to
switch modes back and forth several times in the middle of a session.