use the new preparation mechanism in testing.
The old mecahinsm for doing this setup was extremely fragile in testing.
If other tests ran first, importing all the fundy modules before
globals.setup_for_translation was set to True, then this test would fail
by trying to translate the show method.

The new method is much more robust. However, now running this test
before another test that uses the show statement (somewhat unlikely as
it's gui dependent) will fail. Planning on reorganising the tests so
that a test can be defined once and run on CPython, on CPython prepared
for translation, and on the translated interpreter.