Upgrade the resolution of expression chains.
Since operator properties (precedence, associativity, fixity) are
changeable at runtime, the grammar cannot resolve expression chains.
There is a secondary "parser" inside the asteval module that handles
this, turning the AST for an expression chain into a graph using the
current context. Once an expression has been resolved into a graph the
application of operators/functions to arguments is fully resolved, and
no longer dependent on the state of name bindings.

The old "parser" doing this job was based on the shunting yard
algorithm. The new version is based on top down operator precedence.
Among the improvements available immediately is that operators can now
be used as expressions on their own to refer to the function they are
bound to, e.g. "(+) 1 2" is the same as "1 + 2". Also, the unary
negation operator, though defined, never worked with the old parser.

Things like "(+1)" and "(1+)" seem to work in a limited fashion, but may
not be reliable. At the very least, the supplied argument is always
treated as the left argument, regardless of position.