Add a builtin if function.
No dedicated syntax, so it's just a 3-arg function called if.

Significantly redid the builtin node classes, as manually writing
another version for 3-arg functions would've been a pain in the butt.
Also improved the add_dot_fn method so callables can be passed in the
format spec dictionaries (they are applied to the object being dotted to
generate the string to write in the dot code).